+++ src/numbers/factorization.rs
@@ -7,7 +7,7 @@ use crate::util::{intpow, intpow_const};
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct Factorization<const L: usize> {
     value: u128,
-    prime_powers: [(u128, u128); L],
+    prime_powers: [(u128, usize); L],
 }
 
 
@@ -28,11 +28,11 @@ pub trait Factor<S, const L: usize> {
 
 impl<const L: usize> Factorization<L> {
     /// Creates a new factorization from the given prime powers.
-    pub const fn new(prime_powers: [(u128, u128); L]) -> Factorization<L> {
+    pub const fn new(prime_powers: [(u128, usize); L]) -> Factorization<L> {
         let mut value = 1;
         let mut i = 0;
         while i < L {
-            value *= intpow_const::<0>(prime_powers[i].0, prime_powers[i].1);
+            value *= intpow_const::<0>(prime_powers[i].0, prime_powers[i].1 as u128);
             i += 1;
         }
         Factorization {
@@ -49,7 +49,7 @@ impl<const L: usize> Factorization<L> {
     /// * there are no elements $k | n$ such that $d | k$ and $k \leq l$.
     ///
     /// The iterator cannot outlive the `Factorization`, although the vectors yielded by it may.
-    pub fn maximal_divisors(&self, l: u128) -> impl Iterator<Item = [u128; L]> + '_ {
+    pub fn maximal_divisors(&self, l: u128) -> impl Iterator<Item = [usize; L]> + '_ {
         FactorStream::new(&self.prime_powers, l, true)
     }
 
@@ -62,12 +62,12 @@ impl<const L: usize> Factorization<L> {
     /// Returns the prime power factor represented by prime number `i`, $p_i^{t_i}$.
     /// This method will `panic` if `i` is out of bounds.
     pub fn factor(&self, i: usize) -> u128 {
-        unsafe { intpow::<0>(self.prime_powers[i].0, self.prime_powers[i].1) }
+        unsafe { intpow::<0>(self.prime_powers[i].0, self.prime_powers[i].1 as u128) }
     }
 
     /// Gets the prime powers as an array.
     /// The first element of each entry is the prime, and the second is the power.
-    pub fn prime_powers(&self) -> &[(u128, u128); L] {
+    pub fn prime_powers(&self) -> &[(u128, usize); L] {
         &self.prime_powers
     }
 
@@ -77,13 +77,13 @@ impl<const L: usize> Factorization<L> {
     }
 
     /// Converts an array of powers on the prime factors into an integer.
-    pub fn from_powers(&self, ds: &[u128; L]) -> u128 {
+    pub fn from_powers(&self, ds: &[usize; L]) -> u128 {
         let mut total = 1;
         for (d, (p, t)) in ds.iter().zip(self.prime_powers) {
             if *d > t {
                 return 0;
             } else {
-                total *= unsafe { intpow::<0>(p, *d) };
+                total *= unsafe { intpow::<0>(p, *d as u128) };
             }
         }
         total
@@ -91,9 +91,9 @@ impl<const L: usize> Factorization<L> {
 }
 
 impl<const L: usize> Index<usize> for Factorization<L> {
-    type Output = (u128, u128);
+    type Output = (u128, usize);
 
-    fn index(&self, index: usize) -> &(u128, u128) {
+    fn index(&self, index: usize) -> &(u128, usize) {
         &self.prime_powers[index]
     }
 }
